[FIX/#78] UpComingBdayView 리펙토링 완료
