+++ DoBday/DoBday/View/SaveBdayView/PhotoPickerForSaveBdayView.swift
@@ -8,9 +8,9 @@ import SwiftUI
 import PhotosUI
 
 struct PhotoPickerForSaveBdayView: View {
+
     @Binding var imageData: Data?
     @Binding var selectedItem: PhotosPickerItem?
-
 
     var body: some View {
         //MARK: 이미지 피커
@@ -44,6 +44,5 @@ struct PhotoPickerForSaveBdayView: View {
                 .frame(width: 30, height: 30)
                 .offset(x: 30, y: 40)
         }
-
     }
 }
