+++ DoBday/DoBday/View/SaveBdayView/SetBirthForSaveBdayView/ButtonForCheckingIsLunar.swift
@@ -40,6 +40,5 @@ struct ButtonForCheckingIsLunar: View {
                 }
             }
         }
-
     }
 }
