+++ DoBday/DoBday/View/SaveBdayView/SetBirthForSaveBdayView/ButtonForCheckingTheDate.swift
@@ -32,6 +32,5 @@ struct ButtonForCheckingTheDate: View {
                     .presentationDetents([.medium])
             }
         }
-
     }
 }
