+++ DoBday/DoBday/View/SetDateView/ButtonOfShowingSheetForSettingDateView.swift
@@ -24,7 +24,6 @@ struct ButtonOfShowingSheetForSettingDateView: View {
                         .foregroundColor(.white)
                 }
             }
-
         }
     }
 }
